The output of the Python code doesn't match that of the C code
I suspect the problem has to do with confusion related to the use
and inversion of the larger_tol and to_upper flags. I've identified
a class of tests which do work (largely, all upper case strings)
and I've started to add functionality to help compare output where
the only difference is inverted flags.

Will sort out the tests at a later stage...