Fixed problem with strcmp.c not recognising upper-case conversion flag
The flag array, ind_c, is declared as an array (of unspecified size) holding
pointers to ints (`int *ind_c[]`). But it is *accessed* as if it just contained
ints (e.g. `!ind_c[1]`, line 149)

This mismatch works if you're trying to access the first flag (ind_c[0]) but not
the second (different size ints (or int pointers) on '95 hardware compared to
todays?)

So I simply redeclared the array in the function signature to be an array of
ints, and that compiles without warning and seems to work for my tests.