Fixed problem with strcmp.c not recognising upper-case conversion flag
The flag array, `ind_c`, is declared as an array (of unspecified size) holding
pointers to ints (`int *ind_c[]`). But it is *accessed* as if it just contained
ints -e.g. `!ind_c[1]`,  [line 149](https://github.com/richmilne/JaroWinkler/blob/43e7e186d791622ba6636402b1c2590ef765cd51/jaro/strcmp95.c#L149).

This mismatch works if you're trying to access the first flag (`ind_c[0]`) but
not the second (differently sized ints (or int pointers) on '95 hardware
compared to today's?)

So I simply re-declared the array in the function signature to be an array of
ints, and that compiles without warning and seems to work for my tests.