Fix problems with v2.0.1 packaging (#5)
Previous packages were produced with `setup.py bdist` - but either I did
something different for the last release, or there's now something different
about the files produced by that option.

Going forward I'll now be producing packages with `setup.py bdist_wheel`, as it
is meant to be a more accommodating, platform-agnostic format.

The changes in this PR have no real bearing on how the package is produced (it's
still done manually); I just took the opportunity to add a few comments.